plugins: feature #449 allow fract DPI values when importing PDF
    This commit fixes the issue when requested page width/height was
    not honored because DPI resolution value was constrained to whole
    integer numbers only.

+++ plug-ins/common/file-pdf-load.c
@@ -85,7 +85,7 @@ static gint32            load_image        (PopplerDocument        *doc,
                                             const gchar            *filename,
                                             GimpRunMode             run_mode,
                                             GimpPageSelectorTarget  target,
-                                            guint32                 resolution,
+                                            gdouble                 resolution,
                                             gboolean                antialias,
                                             PdfSelectedPages       *pages);
 
@@ -1019,7 +1019,7 @@ load_image (PopplerDocument        *doc,
             const gchar            *filename,
             GimpRunMode             run_mode,
             GimpPageSelectorTarget  target,
-            guint32                 resolution,
+            gdouble                 resolution,
             gboolean                antialias,
             PdfSelectedPages       *pages)
 {
